backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
aiofiles>=23.2.1
python-multipart>=0.0.9
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.3.0
python-magic==0.4.27; sys_platform != "win32"
python-magic-bin==0.4.14; sys_platform == "win32"
clamd==1.0.2
alembic>=1.13.2
PyPDF2>=3.0.1
chromadb>=0.5.5
google-generativeai>=0.8.0
google-genai>=0.2.0
greenlet>=3.0.3
PyMuPDF>=1.23.0
Pillow>=10.3.0
pytesseract>=0.3.10
pytest>=8.3.2
pytest-asyncio>=0.23.8
pytest-cov>=5.0.0
httpx>=0.27.0
bcrypt>=4.1.0
PyJWT>=2.8.0
email-validator>=2.1.0
redis>=6.4.0
fsrs>=4.0.0
